blogapp/requirements.txt
Django==5.1
environ==1.0
Pillow==9.4.0
Pillow==10.4.0
